refactor(pallets/communities): split logical modules within pallet
