+++ pallets/communities/src/lib.rs
@@ -200,9 +200,6 @@
 pub use pallet::*;
 
 #[cfg(test)]
-mod mock;
-
-#[cfg(test)]
 mod tests;
 
 #[cfg(feature = "runtime-benchmarks")]
@@ -360,6 +357,8 @@ pub mod pallet {
 	// weight and must return a DispatchResult.
 	#[pallet::call]
 	impl<T: Config> Pallet<T> {
+		// === Registry management ===
+
 		/// Registers an appliation as a new community, taking an
 		/// [existential deposit][8] used to create the community account.
 		///
@@ -422,6 +421,8 @@ pub mod pallet {
 			Ok(())
 		}
 
+		// === Memberships management ===
+
 		/// Enroll an account as a community member. In theory,
 		/// any community member should be able to add a member. However, this
 		/// can be changed to ensure it is a privileged function.
@@ -462,6 +463,8 @@ pub mod pallet {
 
 			Ok(())
 		}
+
+		// === Treasury management ===
 
 		/// Transfers an amount of a given asset from the treasury account to a
 		/// beneficiary.
